[cmake] Optimize cmake targets build time
