--- packages.config.in
@@ -1,18 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="boost_program_options-${MSVC_SUFFIX}" version="${BOOST_VERSION}" targetFramework="Native" />
-  <package id="boost_system-${MSVC_SUFFIX}" version="${BOOST_VERSION}" targetFramework="Native" />
-  <package id="boost_filesystem-${MSVC_SUFFIX}" version="${BOOST_VERSION}" targetFramework="Native" />
-  <package id="boost" version="${BOOST_VERSION}" targetFramework="Native" />
-  <package id="zlib" version="1.2.8.8" targetFramework="Native" />
-  <package id="glib" version="2.36.2.11" targetFramework="Native" />
-  <package id="glib.redist" version="2.36.2.11" targetFramework="Native" />
-  <package id="libffi" version="3.0.10.3" targetFramework="Native" />
-  <package id="libffi.redist" version="3.0.10.3" targetFramework="Native" />
-  <package id="libiconv" version="1.14.0.11" targetFramework="Native" />
-  <package id="libiconv.redist" version="1.14.0.11" targetFramework="Native" />
-  <package id="libintl" version="0.18.2.10" targetFramework="Native" />
-  <package id="libintl.redist" version="0.18.2.10" targetFramework="Native" />
-  <package id="zlib" version="1.2.8.8" targetFramework="Native" />
-  <package id="zlib.v120.windesktop.msvcstl.dyn.rt-dyn" version="1.2.8.8" targetFramework="Native" />
-</packages>
