+++ sc-kpm/sc-common/CMakeLists.txt
@@ -1,7 +1,6 @@
 set(SC_COMMON_SRC ${CMAKE_CURRENT_LIST_DIR})
 
 set(SOURCES "sc_keynodes.c")
-
 set(HEADERS "sc_keynodes.h" "sc_common_types.h")
 
 add_library(sc-kpm-common SHARED ${SOURCES} ${HEADERS})
@@ -13,7 +12,7 @@ target_include_directories(sc-kpm-common
 )
 
 add_dependencies(sc-kpm-common sc-core)
-target_link_libraries(sc-kpm-common sc-core ${GLIB2_LIBRARIES})
+target_link_libraries(sc-kpm-common PRIVATE sc-core ${GLIB2_LIBRARIES})
 target_compile_definitions(sc-kpm-common PUBLIC "-DSC_KPM_COMMON_SELF_BUILD")
 
 if(${SC_CLANG_FORMAT_CODE})
