+++ sc-kpm/sc-search/CMakeLists.txt
@@ -8,7 +8,7 @@ target_include_directories(sc-search
     PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}
 )
 add_dependencies(sc-search sc-memory sc-kpm-common)
-target_link_libraries(sc-search sc-kpm-common)
+target_link_libraries(sc-search PRIVATE sc-memory sc-kpm-common)
 
 if(${SC_CLANG_FORMAT_CODE})
     target_clangformat_setup(sc-search)
