+++ sc-kpm/sc-ui/CMakeLists.txt
@@ -7,10 +7,9 @@ target_include_directories(sc-ui
     PRIVATE ${GLIB2_INCLUDE_DIRS}
     PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}
 )
-add_dependencies(sc-ui sc-memory sc-core sc-kpm-common)
 
-find_package(nlohmann_json)
-target_link_libraries(sc-ui sc-memory sc-kpm-common nlohmann_json::nlohmann_json)
+add_dependencies(sc-ui sc-memory sc-core sc-kpm-common)
+target_link_libraries(sc-ui PRIVATE sc-memory sc-kpm-common nlohmann_json::nlohmann_json)
 
 if(${SC_CLANG_FORMAT_CODE})
     target_clangformat_setup(sc-ui)
