+++ sc-kpm/sc-utils/CMakeLists.txt
@@ -7,7 +7,7 @@ target_include_directories(sc-utils
     PRIVATE ${GLIB2_INCLUDE_DIRS}
 )
 add_dependencies(sc-utils sc-core sc-kpm-common sc-search)
-target_link_libraries(sc-utils sc-kpm-common sc-search)
+target_link_libraries(sc-utils PRIVATE sc-kpm-common sc-search)
 
 if(${SC_CLANG_FORMAT_CODE})
     target_clangformat_setup(sc-utils)
