+++ sc-memory/sc-memory/CMakeLists.txt
@@ -10,8 +10,8 @@ add_custom_command(OUTPUT "${CMAKE_CURRENT_LIST_DIR}/scs/scsLexer.cpp"
 )
 
 add_custom_target(SCsParser ALL
-      DEPENDS "${CMAKE_CURRENT_LIST_DIR}"
-      SOURCES "scs/scs.g4"
+    DEPENDS "${CMAKE_CURRENT_LIST_DIR}"
+    SOURCES "scs/scs.g4"
 )
 
 set(SOURCES_PARSER
@@ -30,31 +30,14 @@ target_include_directories(sc-memory
     PRIVATE ${SC_MACHINE_ANTRL_DIR}/runtime/src
     PRIVATE ${SC_MEMORY_SRC}
 )
-
-if(${WIN32})
-    target_link_libraries(sc-memory
-        sc-core
-        ${LIBCURL_LIBRARIES}
-        antlr4_static
-    )
-elseif(${UNIX})
-    target_link_libraries(sc-memory
-        sc-core
-        antlr4_static
-        ${LIBCURL_LIBRARIES}
-        ${Boost_LIBRARIES}
-    )
-endif(${WIN32})
-
-sc_codegen_ex(sc-memory ${CMAKE_CURRENT_SOURCE_DIR} "${CMAKE_CURRENT_SOURCE_DIR}/generated")
-add_dependencies(sc-memory 
+target_link_libraries(sc-memory sc-core antlr4_static)
+add_dependencies(sc-memory
     sc-core
-    sc-code-generator
     antlr4_static
     SCsParser
 )
 
-target_compile_definitions(sc-memory PRIVATE "-DSC_MEMORY_SELF_BUILD")
+sc_codegen_ex(sc-memory ${CMAKE_CURRENT_SOURCE_DIR} "${CMAKE_CURRENT_SOURCE_DIR}/generated")
 
 if(${SC_CLANG_FORMAT_CODE})
     target_clangformat_setup(sc-memory)
