+++ sc-tools/sc-builder/src/scs_loader.hpp
@@ -4,7 +4,6 @@
  * (See accompanying file COPYING.MIT or copy at http://opensource.org/licenses/MIT)
  */
 
-#include "utils.hpp"
 #include "sc-memory/sc_memory.hpp"
 
 class ScsLoader
