+++ sc-tools/sc-server/src/sc-server-impl/sc-memory-json/sc-memory-json-action/sc_memory_check_elements_json_action.hpp
@@ -12,21 +12,7 @@ class ScMemoryCheckElementsJsonAction : public ScMemoryJsonAction
 {
 public:
   ScMemoryJsonPayload Complete(
-      ScMemoryContext * context, ScMemoryJsonPayload requestPayload, ScMemoryJsonPayload & errorsPayload) override
-  {
-    ScMemoryJsonPayload responsePayload;
-
-    for (auto & hash : requestPayload)
-    {
-      ScAddr const addr{hash.get<size_t>()};
-
-      size_t type = 0;
-      if (addr.IsValid())
-        type = context->GetElementType(addr);
-
-      responsePayload.push_back(type);
-    }
-
-    return responsePayload;
-  }
+      ScMemoryContext * context,
+      ScMemoryJsonPayload requestPayload,
+      ScMemoryJsonPayload & errorsPayload) override;
 };
