+++ sc-tools/sc-server/src/sc-server-impl/sc-memory-json/sc-memory-json-action/sc_memory_create_elements_by_scs_json_action.hpp
@@ -4,12 +4,9 @@
  * (See accompanying file COPYING.MIT or copy at http://opensource.org/licenses/MIT)
  */
 
-
 #pragma once
 
 #include "sc_memory_json_action.hpp"
-
-#include "sc-memory/sc_link.hpp"
 
 #include "sc-memory/sc_scs_helper.hpp"
 
@@ -26,50 +23,9 @@ class ScMemoryCreateElementsByScsJsonAction : public ScMemoryJsonAction
 {
 public:
   ScMemoryJsonPayload Complete(
-      ScMemoryContext * context, ScMemoryJsonPayload requestPayload, ScMemoryJsonPayload & errorsPayload) override
-  {
-    ScMemoryJsonPayload responsePayload = ScMemoryJsonPayload::array({});
-
-    sc_uint32 i = 0;
-    for (auto & atom : requestPayload)
-    {
-      std::string scs;
-      ScAddr outputStructure;
-      if (atom.is_string())
-      {
-        scs = atom.get<std::string>();
-      }
-      else
-      {
-        scs = atom["scs"].get<std::string>();
-        outputStructure = ScAddr(atom["output_structure"].get<size_t>());
-      }
-
-      if (m_helper == nullptr)
-        m_helper = new SCsHelper{*context, std::make_shared<DummyFileInterface>()};
-
-      sc_bool textGenResult = SC_FALSE;
-
-      try
-      {
-         m_helper->GenerateBySCsTextLazy(scs, outputStructure);
-         textGenResult = SC_TRUE;
-      }
-      catch (utils::ScException const & e)
-      {
-        SC_LOG_ERROR(e.Message());
-        errorsPayload.push_back({{"ref", i}, {"message", e.Message()}});
-      }
-
-      responsePayload.push_back(textGenResult);
-      ++i;
-    }
-
-    if (responsePayload.is_null())
-      return "{}"_json;
-
-    return responsePayload;
-  }
+      ScMemoryContext * context,
+      ScMemoryJsonPayload requestPayload,
+      ScMemoryJsonPayload & errorsPayload) override;
 
   ~ScMemoryCreateElementsByScsJsonAction() override
   {
