+++ sc-tools/sc-server/src/sc-server-impl/sc-memory-json/sc-memory-json-action/sc_memory_handle_keynodes_json_action.hpp
@@ -12,30 +12,7 @@ class ScMemoryHandleKeynodesJsonAction : public ScMemoryJsonAction
 {
 public:
   ScMemoryJsonPayload Complete(
-      ScMemoryContext * context, ScMemoryJsonPayload requestPayload, ScMemoryJsonPayload & statusPayload) override
-  {
-    ScMemoryJsonPayload responsePayload;
-
-    for (auto & atom : requestPayload)
-    {
-      std::string const & idtf = atom["idtf"].get<std::string>();
-      std::string const & type = atom["command"].get<std::string>();
-
-      ScAddr keynode;
-      if (type == "find")
-        keynode = context->HelperFindBySystemIdtf(idtf);
-      else if (type == "resolve")
-      {
-        ScType const & elType = ScType(atom["elType"].get<size_t>());
-        keynode = context->HelperResolveSystemIdtf(idtf, elType);
-      }
-
-      responsePayload.push_back(keynode.Hash());
-    }
-
-    if (responsePayload.is_null())
-      return "{}"_json;
-
-    return responsePayload;
-  }
+      ScMemoryContext * context,
+      ScMemoryJsonPayload requestPayload,
+      ScMemoryJsonPayload & statusPayload) override;
 };
