+++ sc-tools/sc-server/src/sc-server-impl/sc-memory-json/sc-memory-json-action/sc_memory_json_action.hpp
@@ -13,7 +13,9 @@ class ScMemoryJsonAction
 {
 public:
   virtual ScMemoryJsonPayload Complete(
-      ScMemoryContext * context, ScMemoryJsonPayload requestPayload, ScMemoryJsonPayload & errorsPayload) = 0;
+      ScMemoryContext * context,
+      ScMemoryJsonPayload requestPayload,
+      ScMemoryJsonPayload & errorsPayload) = 0;
 
   virtual ~ScMemoryJsonAction() = default;
 };
