+++ sc-tools/sc-server/src/sc-server-impl/sc-memory-json/sc-memory-json-action/sc_memory_template_generate_json_action.hpp
@@ -12,18 +12,7 @@ class ScMemoryTemplateGenerateJsonAction : public ScMemoryMakeTemplateJsonAction
 {
 public:
   ScMemoryJsonPayload Complete(
-      ScMemoryContext * context, ScMemoryJsonPayload requestPayload, ScMemoryJsonPayload & errorsPayload) override
-  {
-    ScTemplateGenResult result;
-    auto pair = GetTemplate(context, requestPayload);
-    context->HelperGenTemplate(*pair.first, result, pair.second);
-
-    std::vector<size_t> hashesVectors;
-    for (size_t i = 0; i < result.Size(); ++i)
-      hashesVectors.push_back(result[i].Hash());
-
-    ScMemoryJsonPayload const & resultPayload = {{"aliases", result.GetReplacements()}, {"addrs", hashesVectors}};
-    delete pair.first;
-    return resultPayload;
-  }
+      ScMemoryContext * context,
+      ScMemoryJsonPayload requestPayload,
+      ScMemoryJsonPayload & errorsPayload) override;
 };
