+++ sc-tools/sc-server/src/sc-server-impl/sc-memory-json/sc-memory-json-event/sc_memory_json_events_handler.hpp
@@ -47,8 +47,12 @@ private:
       sc_bool & isEvent) override;
 
   ScMemoryJsonPayload HandleCreate(
-      ScServerConnectionHandle const & hdl, ScMemoryJsonPayload const & message, ScMemoryJsonPayload & errorsPayload);
+      ScServerConnectionHandle const & hdl,
+      ScMemoryJsonPayload const & message,
+      ScMemoryJsonPayload & errorsPayload);
 
   ScMemoryJsonPayload HandleDelete(
-      ScServerConnectionHandle const & hdl, ScMemoryJsonPayload const & message, ScMemoryJsonPayload & errorsPayload);
+      ScServerConnectionHandle const & hdl,
+      ScMemoryJsonPayload const & message,
+      ScMemoryJsonPayload & errorsPayload);
 };
