+++ sc-tools/sc-server/src/sc-server-impl/sc_server_log.hpp
@@ -14,7 +14,10 @@ class ScServerLog
 {
 public:
   ScServerLog(
-      ScServerCore * server, std::string const & logType, std::string const & logFile, std::string const & logLevel)
+      ScServerCore * server,
+      std::string const & logType,
+      std::string const & logFile,
+      std::string const & logLevel)
     : m_server(server)
   {
     Initialize(logType, logFile, logLevel);
